Set up test schema / recordbatch generator target
