+++ codegen/fletchgen/CMakeLists.txt
@@ -20,20 +20,20 @@ set(SOURCE_FILES
         src/logging.h
         src/meta.h
         src/constants.h
-        src/common.h                src/common.cc
-        src/arrow-utils.h           src/arrow-utils.cc
-        src/printers.h              src/printers.cc
+        src/common.h src/common.cc
+        src/arrow-utils.h src/arrow-utils.cc
+        src/printers.h src/printers.cc
 
-        src/vhdl/vhdl.h             src/vhdl/vhdl.cc
-        src/stream.h                src/stream.cc
+        src/vhdl/vhdl.h src/vhdl/vhdl.cc
+        src/stream.h src/stream.cc
 
-        src/column.h                src/column.cc
-        src/arbiter.h               src/arbiter.cc
-        src/fletcher-streams.h      src/fletcher-streams.cc
-        src/fletcher-ports.h        src/fletcher-ports.cc
-        src/column-wrapper.h        src/column-wrapper.cc
-        src/usercore.h              src/usercore.cc
-        src/usercore-controller.h   src/usercore-controller.cc
+        src/column.h src/column.cc
+        src/arbiter.h src/arbiter.cc
+        src/fletcher-streams.h src/fletcher-streams.cc
+        src/fletcher-ports.h src/fletcher-ports.cc
+        src/column-wrapper.h src/column-wrapper.cc
+        src/usercore.h src/usercore.cc
+        src/usercore-controller.h src/usercore-controller.cc
 
         src/srec/srec.h src/srec/srec.cc
         src/srec/recordbatch.h src/srec/recordbatch.cc
@@ -41,19 +41,31 @@ set(SOURCE_FILES
         src/vhdt/vhdt.h src/vhdt/vhdt.cc
         src/top/axi.h src/top/axi.cc
         src/top/sim.h src/top/sim.cc
-        src/schema_test.h           src/schema_test.cc
 
-        src/config.h                src/config.cc
+        src/config.h src/config.cc
 
-        src/fletchgen.cc )
+        src/fletchgen.cc)
 
 add_executable(${PROJECT_NAME} ${SOURCE_FILES})
 
 find_library(LIB_ARROW arrow)
-
 find_package(Boost COMPONENTS program_options REQUIRED)
-
 target_link_libraries(${PROJECT_NAME} ${REQUIRED} Boost::program_options)
 target_link_libraries(${PROJECT_NAME} ${REQUIRED} ${LIB_ARROW})
 
 install(TARGETS ${PROJECT_NAME} DESTINATION bin)
+
+option(ENABLE_TESTS "Enable building of test stuff" ON)
+
+message("Building tests.")
+
+if (ENABLE_TESTS)
+    set(TEST_NAME ${PROJECT_NAME}-test-generate)
+    set(TEST_SOURCE_FILES
+            src/arrow-utils.h src/arrow-utils.cc
+            test/test_schemas.h test/test_schemas.cc
+            test/test-generate.cc)
+    add_executable(${PROJECT_NAME}-test-generate ${TEST_SOURCE_FILES})
+    target_link_libraries(${TEST_NAME} ${REQUIRED} Boost::program_options)
+    target_link_libraries(${TEST_NAME} ${REQUIRED} ${LIB_ARROW})
+endif (ENABLE_TESTS)
